RE-17 ->merge-> Dev->QA

+++ database/migrations/2020_06_04_115545_create_sigecig_recibo_cheque_table.php
@@ -16,7 +16,7 @@ class CreateSigecigReciboChequeTable extends Migration
         Schema::create('sigecig_recibo_cheque', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('numero_recibo');
-            $table->bigInteger('numero_cheque');
+            $table->integer('numero_cheque');
             $table->double('monto',10,2);
             $table->string('nombre_banco')->nullable();
             $table->integer('usuario_id');
